requirements.txt
kivymd==1.2.0
plyer==2.1.0
pillow==10.3.0
python-dotenv==1.0.1
